requirements.txt
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0

django-filter==25.1